Move queryConfigurationEntry null check out from datasource check
This has done due to an issue occurred in WSO2 Stream Processor HA scenario where database connection has disconnected right after creating the datasource service and before initializing queryConfigurationEntry.
in such a scenario when retrying to create the connection, queryConfigurationEntry setting up part will be ignored since the dataSource is not 'null'. This will lead to a NullPointerException when trying to create a table.